training: introduced record reader & moved preprocessors
