+++ tests/modelling/test_erc_emb_visual.py
@@ -12,7 +12,7 @@ from hlm12erc.modelling.erc_emb_visual import (
     ERCVisualEmbeddingType,
 )
 from hlm12erc.training.erc_data_collator import ERCDataCollator
-from hlm12erc.training.meld_preprocessor_visual import MeldVisualPreprocessor
+from hlm12erc.training.meld_record_preprocessor_visual import MeldVisualPreprocessor
 
 
 class TestERCResNet50VisualEmbeddings(unittest.TestCase):
