+++ tests/training/test_meld_record_preprocessor_audio.py
@@ -3,7 +3,7 @@ import unittest
 import wave
 
 # My Packages and Modules
-from hlm12erc.training.meld_preprocessor_audio import MeldAudioPreprocessor
+from hlm12erc.training.meld_record_preprocessor_audio import MeldAudioPreprocessor
 
 
 class TestMeldAudioPreprocessor(unittest.TestCase):
